+++ vip-integrations/vip-security-boost-0.1/utils/class-tracking.php
@@ -6,7 +6,7 @@
 namespace Automattic\VIP\Security\Utils;
 
 use Automattic\VIP\Security\Constants;
-use Automattic\VIP\Telemetry\Tracks;
+use Automattic\VIP\Telemetry\Telemetry;
 
 class Tracking {
 
@@ -16,9 +16,9 @@ class Tracking {
 	const PREFIX = 'vip_security_boost';
 
 	/**
-	 * Tracks instance.
+	 * Telemetry instance.
 	 */
-	private static ?Tracks $tracks = null;
+	private static ?Telemetry $telemetry = null;
 
 	private Counter $mfa_display_counter;
 	private Counter $mfa_filter_click_counter;
@@ -28,23 +28,55 @@ class Tracking {
 	private Counter $privileged_email_sent_counter;
 
 	/**
-	 * Get the Tracks instance.
+	 * Get the Telemetry instance.
 	 *
-	 * @return Tracks|null The Tracks instance or null if not available.
+	 * @return Telemetry|null The Telemetry instance or null if not available.
 	 */
-	private static function get_tracks(): ?Tracks {
-		if ( null === self::$tracks && class_exists( '\Automattic\VIP\Telemetry\Tracks' ) ) {
-			self::$tracks = new Tracks(
-				self::PREFIX . '_',
+	private static function get_telemetry(): ?Telemetry {
+		if ( null === self::$telemetry && class_exists( '\Automattic\VIP\Telemetry\Telemetry' ) ) {
+			self::$telemetry = new Telemetry(
+				self::PREFIX . '_' . self::maybe_get_non_production_prefix(),
 				[
 					'plugin_name' => Constants::LOG_PLUGIN_NAME,
 					'site_id'     => defined( 'VIP_GO_APP_ID' ) ? VIP_GO_APP_ID : 0,
 				]
 			);
 		}
-		return self::$tracks;
+		return self::$telemetry;
 	}
 
+	/**
+	 * Record an event using Telemetry, only if Telemetry class is available.
+	 *
+	 * @param string $event_name Event name.
+	 * @param array  $event_data Event data.
+	 */
+	private static function record_event( $event_name, $event_data = [] ) {
+		$telemetry = self::get_telemetry();
+		if ( $telemetry ) {
+			$telemetry->record_event( $event_name, $event_data );
+		}
+	}
+
+	/**
+	 * Get the prefix for stats and events.
+	 *
+	 * @return string `local_` if local, `nonprod_` if nonproduction, empty otherwise.
+	 */
+	private static function maybe_get_non_production_prefix( $trailing_underscore = true ): string {
+		$trailing_underscore = $trailing_underscore ? '_' : '';
+		if ( ! defined( 'VIP_GO_APP_ENVIRONMENT' ) || 'local' === constant( 'VIP_GO_APP_ENVIRONMENT' ) ) {
+			return 'local' . $trailing_underscore;
+		}
+		if ( 'production' !== constant( 'VIP_GO_APP_ENVIRONMENT' ) ) {
+			return 'nonprod' . $trailing_underscore;
+		}
+		return '';
+	}
+
+	/**
+	 * Initialize stats used by the class-collector.php for the prometheus stats.
+	 */
 	public function initialize( RegistryInterface $registry ): void {
 		$this->mfa_display_counter = $registry->getOrRegisterCounter(
 			'vip_security_boost',
@@ -90,59 +122,42 @@ class Tracking {
 	}
 
 	public static function mfa_display( $filter_enabled ) {
-		$tracks = self::get_tracks();
-		if ( $tracks ) {
-			$tracks->record_event( 'mfa_page_view', [ 'filtered' => $filter_enabled ] );
-		}
+		self::record_event( 'mfa_page_view', [ 'filtered' => $filter_enabled ] );
 		self::record_stats( 'mfa-display' . ( $filter_enabled ? '-filtered' : '' ) );
 	}
 
 	public static function mfa_filter_click( $filter_type ) {
-		$tracks = self::get_tracks();
-		if ( $tracks ) {
-			$tracks->record_event( 'mfa_filter_click', [ 'filter_type' => $filter_type ] );
-		}
+		self::record_event( 'mfa_filter_click', [ 'filter_type' => $filter_type ] );
 		self::record_stats( 'mfa-filter-click' );
 	}
 
 	public static function mfa_sorting( $sort_column, $sort_order ) {
-		$tracks = self::get_tracks();
-		if ( $tracks ) {
-			$tracks->record_event( 'mfa_sort', [
-				'sort_column' => $sort_column,
-				'sort_order'  => $sort_order,
-			] );
-		}
+		self::record_event( 'mfa_sort', [
+			'sort_column' => $sort_column,
+			'sort_order'  => $sort_order,
+		] );
 		self::record_stats( 'mfa-sorting' );
 	}
 
 	public static function blocked_users_view() {
-		$tracks = self::get_tracks();
-		if ( $tracks ) {
-			$tracks->record_event( 'blocked_users_view' );
-		}
+		self::record_event( 'blocked_users_view' );
 		self::record_stats( 'blocked-users-view' );
 	}
 
 	public static function user_unblock( $user_id, $user_role ) {
-		$tracks = self::get_tracks();
-		if ( $tracks ) {
-			$tracks->record_event( 'blocked_users_unblock', [
-				'user_role'   => $user_role,
-				'has_user_id' => ! empty( $user_id ),
-			] );
-		}
+		self::record_event( 'blocked_users_unblock', [
+			'user_role'   => $user_role,
+			'has_user_id' => ! empty( $user_id ),
+		] );
 		self::record_stats( 'user-unblock' );
 	}
 
 	public static function privileged_email_sent( $email_type, $recipient_role ) {
-		$tracks = self::get_tracks();
-		if ( $tracks ) {
-			$tracks->record_event( 'privileged_activity_email_sent', [
-				'email_type'     => $email_type,
-				'recipient_role' => $recipient_role,
-			] );
-		}
+		self::record_event( 'privileged_activity_email_sent', [
+			'email_type'     => $email_type,
+			'recipient_role' => $recipient_role,
+		] );
+
 		self::record_stats( 'privileged-email-sent' );
 
 		Logger::info( 'vip-security-boost', 'Privileged user email notification sent', [
@@ -159,9 +174,14 @@ class Tracking {
 	 * @param mixed  $value Stat value.
 	 */
 	private static function record_stats( $stat_name ) {
+		$env_prefix = self::maybe_get_non_production_prefix( false );
+		$stat_code  = self::PREFIX;
+		if ( ! empty( $env_prefix ) ) {
+			$stat_code = self::PREFIX . '_' . $env_prefix;
+		}
 		// We're tracking the stats in production only
-		if ( 'local' === constant( 'VIP_GO_APP_ENVIRONMENT' ) ) {
-			Logger::info( 'vip-security-boost', 'Bumping stats for /s/' . self::PREFIX . "/{$stat_name}", [
+		if ( ! defined( 'VIP_GO_APP_ENVIRONMENT' ) || 'local' === constant( 'VIP_GO_APP_ENVIRONMENT' ) ) {
+			Logger::info( 'vip-security-boost', 'Bumping stats for /s/' . $stat_code . '/' . $stat_name, [
 				'stat_name' => $stat_name,
 			] );
 			return;
@@ -169,7 +189,7 @@ class Tracking {
 
 		if ( function_exists( '\Automattic\VIP\Stats\send_pixel' ) ) {
 			try {
-				\Automattic\VIP\Stats\send_pixel( [ self::PREFIX => $stat_name ] );
+				\Automattic\VIP\Stats\send_pixel( [ $stat_code => $stat_name ] );
 			} catch ( \Exception $e ) {
 				Logger::error( 'vip-security-boost', 'Stats recording failed', [
 					'stat_name' => $stat_name,