+++ wp-parsely/CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.20.3](https://github.com/Parsely/wp-parsely/compare/3.20.2...3.20.3) - 2025-06-25
+
+### Changed
+
+- Update "Content Helper" to "Content Intelligence" on settings page ([#3489](github.com/Parsely/wp-parsely/pull/3489))
+
+### Fixed
+
+- Update "Stable tag" to "3.20.2", fix README whitespace ([#3486](https://github.com/Parsely/wp-parsely/pull/3486))
+
+## [3.20.2](https://github.com/Parsely/wp-parsely/compare/3.20.1...3.20.2) - 2025-06-24
+
+### Changed
+
+- Change feature name from "Traffic Boost" to "Engagement Boost" ([#3482](https://github.com/Parsely/wp-parsely/pull/3482))
+
 ## [3.20.1](https://github.com/Parsely/wp-parsely/compare/3.20.0...3.20.1) - 2025-06-19
 
 ### Fixed